feat: add cover images
fix: add button labels
